added date and time support for CSVtoSQL.py

+++ CSVtoSQL.py
@@ -3,10 +3,18 @@ import sys
 import argparse
 import os
 from datetime import datetime
+import re
 
 if int(sys.version_info[0]) >= 3:
 	from functools import reduce
+	xrange = range
+else:
+	from six.moves import input
 
+
+# NOTE: THE CURRENT IMPLEMENTATION OF THIS SCRIPT WILL OVERWRITE ANY TABLE WITH THE SAME NAME IN THE DATABASE
+# 
+# MEANING IT WILL DROP THE TABLE WITH THE SAME NAME BEFORE INSERTING!!!!!!!!!!!
 
 parser = argparse.ArgumentParser()
 parser.add_argument("-c", "--csv", type=str, help="set the filepath to the csv file you would like to add to the server")
@@ -18,6 +26,7 @@ csv = None
 database = None
 
 def format(line, coltypes):
+	format.visited += 1
 	formatted = []
 	for (x, y) in zip(line, coltypes):
 		if x == "" or x.upper() == "N/A":
@@ -30,10 +39,46 @@ def format(line, coltypes):
 			formatted.append(int(x))
 		elif 'char' in y.lower():
 			formatted.append(x)
+		elif 'date' in y.lower():
+			y = y.replace("date", '')
+			r = re.compile('(\\(.*\\))', re.IGNORECASE|re.DOTALL).search(y)
+			if r is None or len(r.groups()) > 1:
+				raise(Exception("When using date, you provide a valid input for the formatting with the parentheses:\n%s" % y))
+			accepted = False
+			for i in xrange(len(x), 1, -1):
+				try:
+					formatted.append(datetime.strptime(x[0:i], y[1:-1]))
+					accepted = True
+					if format.visited < 2 and i != len(x):
+						print("Seems like your date element isn't an exact match to your formatter. To drastically speed up this process, please remove any leading/trailing characters")
+						format.visited += 1
+					break
+				except Exception as e:
+					continue
+			if not accepted:
+				raise(Exception("Could not format date. If there are any trailing/leading characters, try removing them. Otherwise, try editing your formatter"))
+		elif 'time' in y.lower():
+			y = y.replace("time", '')
+			r = re.compile('(\\(.*\\))', re.IGNORECASE|re.DOTALL).search(y)
+			if r is None or len(r.groups()) > 1:
+				raise(Exception("When using time, you provide a valid input for the formatting with the parentheses:\n%s" % y))
+			accepted = False
+			for i in xrange(len(x), 1, -1):
+				try:
+					formatted.append(datetime.strptime(x[0:i], y[1:-1]).strftime("%H:%M:%S"))
+					accepted = True
+					if format.visited < 2 and i != len(x):
+						print("Seems like your date element isn't an exact match to your formatter. To drastically speed up this process, please remove any leading/trailing characters")
+						format.visited += 1
+					break
+				except Exception as e:
+					continue
+			if not accepted:
+				raise(Exception("Could not format time. If there are any trailing/leading characters, try removing them. Otherwise, try editing your formatter"))
+			
+
 		else:
 			raise(Exception(y + " type is either not currently supported or is not a datatype"))
-	# print tuple(formatted)
-	# sys.exit()
 	return formatted
 
 def create_table(tableName, cols, fail=False):
@@ -43,17 +88,27 @@ def create_table(tableName, cols, fail=False):
 		print("%s table made" % tableName)
 	except Exception as e:
 		print(e)
-		if not fail:
-			print("Dropping the table, trying again")
-			c.execute("DROP TABLE %s;" % tableName)
-			create_table(tableName, cols, fail=True)
+		overwrite = None
+		while overwrite is None:
+			x = input("Looks like the table already exists. You can choose to overwrite the table or insert to the current table. Overwrite? (y/n)\n").lower() 
+			if "n" == x or "y" == x:
+				overwrite = x
+		if overwrite == 'y':
+			if not fail:
+				
+				print("Dropping the table, trying again")
+				c.execute("DROP TABLE %s;" % tableName)
+				create_table(tableName, cols, fail=True)
+			else:
+				print("Unable to create table")
+				raise(e)
 		else:
-			print("Unable to create table")
-			raise(e)
+			print("Will move on to try and insert the table.")
 
 
 
 if __name__ == "__main__":
+	format.visited = 0
 	print('')
 	
 	tableName = None
@@ -80,7 +135,7 @@ if __name__ == "__main__":
 	else:
 		raise(Exception("Must provide a table name. Use -t"))
 	if args.coltypes != None:
-		coltypes = (args.coltypes).replace("\n", "").replace("\r", "").split(" ")
+		coltypes = list(((args.coltypes).replace("\n", "").replace("\r", "").split(" ")))
 	else:
 		raise(Exception("Must provide a list of column types which correlate to the csv. Use -ct"))
 
@@ -89,7 +144,7 @@ if __name__ == "__main__":
 
 	cols = csv.pop(0)
 
-	
+
 	for c in cols:
 		for a in " ,./<>?;':\"\][{}|!@#$%^&*()_+=-":
 			if a in c:
@@ -99,13 +154,13 @@ if __name__ == "__main__":
 
 	c = database.cursor()
 
-	#
-	cols = reduce(lambda x, y: x + ", " + y, map(lambda a: a[0] + " " + a[1], zip(cols, coltypes)))
+	cols = reduce(lambda x, y: x + ", " + (y.split(" ")[0] + " date" if " date" in y else y.split(" ")[0] + " time" if " time" in y else y), map(lambda a: a[0] + " " + a[1], zip(cols, coltypes)))
 
 	create_table(tableName, cols)
+	
 	csvProcessed = map(lambda line: format(line, coltypes), csv)
 	for line in csvProcessed:
-		insert_command = "INSERT INTO %s VALUES (%s);" % (tableName, reduce(lambda x, y: x + ", " + y, ["?" for i in range(len(coltypes))]))
+		insert_command = "INSERT INTO %s VALUES (%s);" % (tableName, reduce(lambda x, y: x + ", " + y, ["?" for i in xrange(len(coltypes))]))
 		c.execute(insert_command, line)
 
 	print("Everything inserted!")
